requirements.txt
apache-libcloud
pyyaml